pyproject.toml
[project]
name = "NLP-sentiment-analysis"
version = "0.1.0"
description = "A deep learning project for entity-level sentiment analysis of tweets"
authors = [
    {name = "Abe Alcaraz", email = "abe.alcarazm@gmail.com"}
]
requires-python = ">=3.12, <3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "numpy>=1.26.0",
    "pandas>=2.3.0",
    "scikit-learn>=1.6.1",
    "matplotlib>=3.10.3",
    "fastapi[standard]>=0.115.12",
    "uvicorn>=0.34.2",
    "loguru>=0.7.3",
    "ipykernel>=6.30.1",
    "kagglehub>=0.3.12",
    "tensorflow>=2.19.0",
    "mlflow>=3.2.0",
    "spacy>=3.8.7",
    "pip>=25.2",
    "sentence-transformers>=5.1.0",
    "tf-keras>=2.19.0",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "ruff>=0.1.9",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "nbqa>=1.9.1",
    "black>=25.1.0",
    "isort>=6.0.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Ruff settings
[tool.ruff]
indent-width = 4
line-length = 88
target-version = "py38"
lint.select = ["E", "F", "I", "B", "W", "C90"]
lint.ignore = []
lint.isort = { known-first-party = ["src"] }
lint.per-file-ignores = { "__init__.py" = ["F401"] }

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"
line-ending = "lf"
skip-magic-trailing-comma = false

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
